refactor sqlite
